End2End test works with new SCIOND API.
